[jsc#cap-566] Addressed larger comment, split app into logical pieces.
