+++ assessor/lib/colors.rb
@@ -1,0 +1,32 @@
+
+# ......................................................................
+# Colorization support.
+class String
+  def red
+    "\033[0;31m#{self}\033[0m"
+  end
+
+  def green
+    "\033[0;32m#{self}\033[0m"
+  end
+
+  def yellow
+    "\033[0;33m#{self}\033[0m"
+  end
+
+  def blue
+    "\033[0;34m#{self}\033[0m"
+  end
+
+  def magenta
+    "\033[0;35m#{self}\033[0m"
+  end
+
+  def cyan
+    "\033[0;36m#{self}\033[0m"
+  end
+
+  def bold
+    "\033[0;1m#{self}\033[0m"
+  end
+end
